atualizado tabela criado seed para teste e add gem faker

+++ db/migrate/20230518232839_create_payments.rb
@@ -1,7 +1,7 @@
 class CreatePayments < ActiveRecord::Migration[6.1]
   def change
     create_table :payments do |t|
-      t.decimal :value
+      t.float :value
       t.references :client, null: false, foreign_key: true
       t.references :barber, null: false, foreign_key: true
